function_app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
moviepy
openai
Pillow
requests
pandas
google-auth-oauthlib
google-auth
google-api-python-client